+++ lib/repositories/weather_repository.dart
@@ -1,5 +1,5 @@
 import '../core/network/response/result.dart';
-import '../models/forecast.dart';
+import '../features/weather/models/forecast.dart';
 
 /// 天気データを取得するためのインターフェース。
 abstract class WeatherRepository {
