+++ lib/repositories/weather_repository_impl.dart
@@ -3,7 +3,7 @@ import 'package:cloud_functions/cloud_functions.dart';
 import '../core/logger/app_log.dart';
 import '../core/network/api_error.dart';
 import '../core/network/response/result.dart';
-import '../models/forecast.dart';
+import '../features/weather/models/forecast.dart';
 import '../services/i_weather_api_client.dart';
 import 'weather_repository.dart';
